Rename job and documentation
1. Renamed classes and pushed to a new package
2. Added Javadoc documentation

Thank you for code review!
